Added col_start Shortcode; Using Window-Width instead of Device-Resolution
1.) Added [col_start] shortcode. Add this code just before you add the
first column shortcode. This will use the CSS clear:left to make sure,
columns are displayed correctly even if something fancy is happening
above the columns (i had trouble with text + left-aligned pictures
above the columns - with this shortcode it's solved)

2.) On a Mac/PC the columns were always displayed, even if the
browser-window width was very small. That looked ugly so i decided to
change the CSS.
Now, if the window-width is < 800px, no columns are displayed. This
should work for small devices as well as small windows on big screens.
Feel free to fine-tune that, but i find it hard to do that with all
possible column-combinations. In my case i'm only using 3 equal sized
columns so it was easy to disable them at all. Maybe you like to have a
different fallback for 2 equal columns etc.